+++ js/actions/index.js
@@ -1,11 +1,7 @@
-export const add = () => {
-  return {
-    type: 'ADD'
-  }
-}
+export const add = () => ({
+  type: 'ADD'
+})
 
-export const subtract = () => {
-  return {
-    type: 'SUBTRACT'
-  }
-}
+export const subtract = () => ({
+  type: 'SUBTRACT'
+})
