+++ js/index.js
@@ -1,31 +1,11 @@
-import {
-  StackNavigator
-} from 'react-navigation'
-import Home from './screens/home'
-import Play from './screens/play'
-import {
-  AppRegistry
-} from 'react-native'
+import { AppRegistry } from 'react-native'
 import { Provider } from 'react-redux'
 import { createStore } from 'redux'
 import counterApp from './reducers'
 import React, { Component } from 'react'
+import AppWithNavigationState from './navigator'
 
 let store = createStore(counterApp)
-
-const AppRouteConfigs = {
-  Main: { screen: Home },
-  Play: { screen: Play }
-}
-const AppNavigator = StackNavigator(AppRouteConfigs)
-
-class AppWithNavigationState extends Component {
-  render() {
-    return (
-      <AppNavigator />
-    )
-  }
-}
 
 class HomophoneGame extends Component {
   render() {
