+++ js/reducers/counter.js
@@ -4,18 +4,18 @@ const initialState = {
 
 const counter = (state = initialState, action) => {
   switch (action.type) {
-    case 'ADD':
-      return {
-        value: state.value + 1
-      }
+  case 'ADD':
+    return {
+      value: state.value + 1
+    }
 
-    case 'SUBTRACT':
-      return {
-        value: state.value - 1
-      }
+  case 'SUBTRACT':
+    return {
+      value: state.value - 1
+    }
 
-    default:
-      return state
+  default:
+    return state
   }
 }
